Use a date for the PolicyDocument version instead of semver version number. (#25)
Using a semantic version number implies that the
PolicyDocument format and/or semantics may change in a backward-incompatible
way at some point in the future. This must not be allowed to happen. Using a
date more clearly states the intention of "this release has some semantics
which is distinct but compatible with the semantics from other releases".